+++ changelog.md
@@ -32,6 +32,7 @@
 * Loading shipping by weight grid failed under SQL CE
 * #949 Import: ProcessSlugs does not process explicitly specified "SeName", if product name did not change in an update scenario
 * Customer import: Creates customer role duplicates for inserted customers
+* GMC feed does not generate the sale price if the sale price is set for a future date
 
 
 ## SmartStore.NET 2.5